fix(#436 CI): fold serve_routed tests into net_serve.rs to spare a test binary
`cargo test --workspace` failed CI with "No space left on device" on
the test-link step (run 25966809939). Same disk-pressure class as
#457: every lex-runtime integration test compiles into its own
binary, each statically linked against hyper + tokio + sled + polars
+ rusqlite. Each binary is hundreds of MB; the standalone
`net_serve_routed.rs` introduced in the previous commit pushed the
runner past its expanded budget.

The five `serve_routed` test cases now live as `serve_routed_*` in
the existing `net_serve.rs`, reusing its `spawn_lex_server`,
`wait_for_bind`, and `http` helpers. Coverage is identical; the
delta is a deleted test binary.
